Add Platoon and PvP daily drills to Calculator.vue
This commit adds 2 new toggles for Platoon and PvP dailies, along
with icons.

Also fixed an error where withdrawal receipts could be exchanged
weekly for Access Permissions. This was changed to 30 days to be
more accurate.

Finally, some comments outlining future options to add for
reoccurring sources of Collapse Pieces.