Add create-vue-app.yaml
Create a ymal workflow to host the vue app on GitHub webpages.
This ymal will clone the 'calculator_additions' branch of this
repo fork and build the static app with 'npm run build'. This
yaml makes use of setting an env var GITHUB_PAGES to indicate
that the app is being built for GitHub webpages.