fix rendering dynamicly generated src routes
When creating a static build of the vue app
with 'npm run build', the built site has some
assets that don't load properly when setting
the base URL to something that isn't the root.

This commit fixes the broken routes for dynamicly
generated srcs in vues by adding appending a `./`
to all src tags routes that don't compile to the
page with the proper base URL. This changes the
URL routing from absolute to relative so that
they are processed through the webpack's dependency
graph and resolved as module dependencies.